outgoing-request and -response: rename `write` to `body`, and fix docs
`body` is a better name than `write` - it makes it explicit this is
about access to the body. the `outgoing-body` resource itself has a
method `write` which gives the body contents as a stream. Changing this
method name makes this distinction more clear.

The docs for `outgoing-request.write` were incorrect - the outgoing-body
is not a child of the outgoing-request (or outgoing-response), otherwise
outgoing body streaming (i.e. writing to the body after the request or
response has been initiated) would not be possible.