[DO NOT MERGE] first draft of wasi:http@0.3.0
This is intended to start a conversation about what we expect wasi:http@0.3.0 to
look like once we have support for `stream`s, `future`s, stream `error`s, and
asynchronous lifts and lowers in the component model.  In addition, I'm planning
to implement a prototype of this interface using
[isyswasfa](https://github.com/dicej/isyswasfa) in the near future, so I want to
make sure the design is at least plausible.

High level description:

- The `incoming-handler` and `outgoing-handler` interfaces have been combined
  into a single `handler` interface.

- The `incoming-` and `outgoing-` variations of `request` and `response` have
  been combined.  Likewise for `body`.

- I've added a `option<request-options>` field to `request` since it would be
  awkward to leave it as a parameter of `handler.handle` (e.g. what would it
  mean to receive such a parameter for an incoming request or for a request
  passed from one component to the other without any use of the network?).

- I've added a `request-options-error` (analogous to `header-error`) to
  distinguish between unsupported fields and immutable handles.

I do not intend for this to be merged into the `main` branch of the `wasi-http`
repo (since we'll presumably be doing 0.2.x work there), but I _would_ like for
it to live somewhere we can iterate on it.  A few options come to mind:

- Keep it in my fork of the `wasi-http` repo for now
- Make a long-lived branch in the official repo for it
- Put it in another repo under the `WebAssembly` org (e.g. `wasi-http-0.3.0`)
- Other ideas?

Signed-off-by: Joel Dice <joel.dice@fermyon.com>