Add `http-response-status-invalid` error code
I failed to notice this was missing in #52. Status codes outside the range of 100-599 are
[invalid](https://www.rfc-editor.org/rfc/rfc9110#section-15-6), though there is a nod to
implementation-defined behavior outside those valid ranges. In practice we have found it very useful
to return error codes specific to status problems to customers since our platform does not support
codes outside this range, and the next-best option in the RFC 9209 equivalents is the generic
`http-protocol-error` which doesn't help a user much with debugging.