feat: allow independent retrieval of trailers
This commit changes wasi:http/types#body.finish to forward the
optionally present trailers, rather than retrieve them with the
possibility of generating an error along the way.

With this change, callers of `body.finish` will have to retreive
trailers on their own, which *may* produce an efficiency gain in the
case where trailers are present but not needed, and avoiding work of
checking for trailers in the first place.

Signed-off-by: Victor Adossi <vadossi@cosmonic.com>