[0.3.0-draft] Add an "invalid status" error code.
Some host environments have the ability to report an invalid status
error when a response is received, and there doesn't currently appear
to be a way to translate this other than using `internal-error`.
Since `error-code` has codes like `HTTP-request-method-invalid`, and
`HTTP-request-URI-invalid`, it seems to make sense to add a
`HTTP-response-status-invalid` too.